+++ packages/PowerGrid/src/package.ts
@@ -6,6 +6,7 @@ import {_BinaryImageCellRenderer} from './package.g';
 import {_ImageCellRenderer} from './package.g';
 import {_StarsCellRenderer} from './package.g';
 import {_ColorCellRenderer} from './package.g';
+import {_SvgCellRenderer} from './package.g';
 /* Do not change these import lines to match external modules in webpack configuration */
 import * as grok from 'datagrok-api/grok';
 import * as ui from 'datagrok-api/ui';
@@ -475,3 +476,4 @@ export {_ScatterPlotCellRenderer};
 export {_MultiChoiceCellRenderer};
 export {_StarsCellRenderer};
 export {_ColorCellRenderer};
+export {_SvgCellRenderer};